Migrating FullField from DataTable to separate component
