+++ src/js/components/pipelines/DataTable.jsx
@@ -12,7 +12,8 @@ var d3 = require('d3'),
     dsUtil = require('../../util/dataset-utils'),
     assets = require('../../util/assets'),
     Icon = require('../Icon'),
-    bindChannel = require('../../actions/bindChannel');
+    bindChannel = require('../../actions/bindChannel'),
+    FullField = require('./FullField');
 
 function mapStateToProps(state, ownProps) {
   return {
@@ -66,15 +67,12 @@ var DataTable = React.createClass({
   },
 
   showFullField: function(evt) {
-    var target = evt.target,
-        name = target.textContent,
-        schema = dsUtil.schema(this.props.id);
-
-    this.hideFull(evt);
-    this.setState({fullField: schema[name]});
-    this.$fullField.style('display', 'block')
-      .style('top', target.offsetTop);
-  },
+   var fullField = FullField;
+   this.hideFull(evt);
+   this.setState({fullField: fullField});
+   this.$fullField.style('display', 'block')
+     .style('top', target.offsetTop);
+ },
 
   showFullValue: function(evt) {
     var target = d3.select(evt.target),
@@ -97,71 +95,6 @@ var DataTable = React.createClass({
     this.setState({fullField: null, fullValue: null});
     this.$fullField.style('display', 'none');
     this.$fullValue.style('display', 'none');
-  },
-
-  handleDragStart: function(evt) {
-    this.setState({bindField: this.state.fullField});
-    evt.dataTransfer.setData('text/plain', evt.target.id);
-    evt.dataTransfer.effectAllowed = 'link';
-    sg.set(sg.MODE, 'channels');
-    ctrl.update();
-  },
-
-  handleDragOver: function(evt) {
-    if (evt.preventDefault) {
-      evt.preventDefault(); // Necessary. Allows us to drop.
-    }
-
-    return false;
-  },
-
-  // This makes use of the bubble cursor, which corresponds to the cell signal;
-  // we're using that to figure out which channel we are closest to. The
-  // SELECTED signal indicates the mark to bind the data to.
-  handleDragEnd: function(evt) {
-    var props = this.props,
-        sel = sg.get(sg.SELECTED),
-        cell = sg.get(sg.CELL),
-        bindField = this.state.bindField,
-        dropped = sel._id && cell._id;
-
-    try {
-      if (dropped) {
-        props.bindChannel(props.id, bindField, sel.mark.def.lyra_id, cell.key);
-      }
-    } catch (e) {
-      console.warn('Unable to bind primitive');
-      console.warn(e);
-    }
-
-    sg.set(sg.MODE, 'handles');
-    sg.set(sg.CELL, {});
-    this.setState({bindField: null});
-
-    if (dropped) {
-      this.requestVegaReparse();
-    } else {
-      ctrl.update();
-    }
-  },
-
-  handleDrop: function(evt) {
-    if (evt.preventDefault) {
-      evt.preventDefault(); // Necessary. Allows us to drop.
-    }
-
-    return false;
-  },
-
-  changeMType: function(evt) {
-    var MTYPES = dsUtil.MTYPES,
-        fullField  = this.state.fullField,
-        mTypeIndex = MTYPES.indexOf(fullField.mtype);
-
-    mTypeIndex = (mTypeIndex + 1) % MTYPES.length;
-    fullField.mtype = MTYPES[mTypeIndex];
-
-    this.setState({fullField: fullField});
   },
 
   render: function() {
@@ -192,13 +125,6 @@ var DataTable = React.createClass({
           />
         ) : null;
 
-    fullField = fullField ? (
-      <span>
-        <Icon onClick={this.changeMType}
-          glyph={assets[fullField.mtype]} width="10" height="10" /> {fullField.name}
-      </span>
-      ) : null;
-
     return (
       <div>
         <div className="datatable"
@@ -220,12 +146,8 @@ var DataTable = React.createClass({
             }, this)}
           </tbody></table>
 
-          <div className={'full field ' + props.className}
-            draggable={true}
-            onDragStart={this.handleDragStart}
-            onDragOver={this.handleDragOver}
-            onDragEnd={this.handleDragEnd}
-            onDrop={this.handleDrop}>{fullField}</div>
+
+      <FullField id={this.props.id} className={this.props.className}></FullField>
 
           <div className="full value">{fullValue}</div>
         </div>
